requirements.txt
bcrypt==4.1.2
python-multipart==0.0.6
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
pydantic==2.5.0
python-dotenv==1.0.0
slowapi==0.1.9
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
PyMuPDF==1.23.3
python-jose==3.3.0
passlib==1.7.4
mysql-connector-python==8.3.0
gunicorn==22.0.0
pytest==8.3.5
